requirements.txt
asgiref~=3.3.4
boto3~=1.17.62
botocore~=1.20.62
certifi~=2020.12.5
chardet~=4.0.0
coreapi~=2.3.3
coreschema~=0.0.4
Django~=3.2
django-anymail~=8.3
django-cleanup~=5.2.0
django-cors-headers~=3.7.0
django-phonenumber-field~=5.1.0
django-rest-fuzzysearch
django-rest-registration~=0.6.1
django-storages~=1.11.1
djangorestframework~=3.12.4
djangorestframework-simplejwt~=4.6.0
drf-yasg~=1.20.0
gunicorn~=20.1.0
idna~=2.10
inflection~=0.5.1
itypes~=1.2.0
Jinja2~=2.11.3
jmespath~=0.10.0
mailjet-rest~=1.3.4
MarkupSafe~=1.1.1
openapi-codec~=1.3.2
packaging~=20.9
phonenumbers~=8.12.21
Pillow~=8.2.0
psycopg2-binary~=2.8.6
PyJWT~=2.0.1
pyparsing~=2.4.7
python-dateutil~=2.8.1
python-dotenv~=0.17.0
pytz~=2021.1
requests~=2.25.1
ruamel.yaml~=0.17.4
ruamel.yaml.clib~=0.2.2
s3transfer~=0.4.2
schedule~=1.1.0
simplejson~=3.17.2
six~=1.15.0
sqlparse~=0.4.1
transliterate
typing-extensions~=3.7.4.3
uritemplate~=3.0.1
urllib3~=1.26.4